rename session account middleware to sessions, plural, prep for fixing #6678
